Data type and UI changed

+++ app/src/main/java/com/example/currencyflow/data/CurrencyViewModel.kt
@@ -5,15 +5,16 @@ import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
+import java.math.BigDecimal
 
 class CurrencyViewModel : ViewModel() {
 
     // StateFlow do przechowywania listy walut
-    private val _currencyRates = MutableStateFlow<Map<String, Double>>(emptyMap())
-    val currencyRates: StateFlow<Map<String, Double>> get() = _currencyRates
+    private val _currencyRates = MutableStateFlow<Map<String, BigDecimal>>(emptyMap())
+    val currencyRates: StateFlow<Map<String, BigDecimal>> get() = _currencyRates
 
     // Funkcja do aktualizacji kursów walut
-    fun updateCurrencyRates(newRates: Map<String, Double>) {
+    fun updateCurrencyRates(newRates: Map<String, BigDecimal>) {
         viewModelScope.launch {
             _currencyRates.value = newRates
         }
